11/14 - 1st Build, Some Tweaks & Fixes
Nghi, your jump animation is bugged in that the fall plays on jump and jump only plays once you land, not sure why.

Cole, two clicks are required to initially start your railcar. Please fix this, discard the part where the player moves with it at least for now, and try to add visual to the line connecting both points.